requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
requests==2.32.3
requests-file==2.1.0
setuptools==75.1.0
sqlparse==0.5.1
tldextract==5.1.2
tzdata==2024.2
urllib3==2.2.3
zope.event==5.0
zope.interface==7.0.3
